All: Trash original idea and move to an IR
Without an intermediate representation we're very well screwed. The idea
that output code should be easily human readable wasn't really flexible
and would have costed a lot of work. Now the IR is similar to a compiler
intermediate code with consists of very simple statements.

This basically breaks everything that was working so far. Have a little
faith. Disable the build test since it won't pass