All: re-structure almost everything
Add an intermediate representation which is language independent and
restore the C language definition that we will translate to, and only
later will it get "rendered" to actual C code. This is a mess for now,
but in the long run it should mean the project is more maintainable.

This breaks all the source generation support, but hey, at least the
tests are passing.