Translate: always save return values in lvalues
This way we can free stuff before returning. Using tr_save is not ok
since we want to save it even if the return value is an integer or
something since it may be:

	a = <>
	return length(a)

Where we need to free a before returning.

Check if the expression is already an LValue and do nothing if so