Translate: copy arguments into local variables
To properly keep track of their liveness, since arguments should never
be freed (since they belong to the calling invocation) but if we modify
them then we do need to free the resulting value.

Adding a flag depending on if the argument was written to or not is not
possible because of idempotency and etc. I hope that we can later
optimize most of these copies